mem: Move pruner cursor helpers into the maintenance module
Address review feedback: keep the limiter's cursor surface minimal by
moving the three pruning helpers (has_active_reads, reset_one_idle_cursor,
clear_one_seek_window) out of MemoryLimiter and into the maintenance
module, where they operate on a single per-round live-cursor snapshot.

The limiter now exposes only safe, narrow access to cursor state:
live_cursors() for the snapshot, plus CursorState::{has_active_read,
idle_tick, clear_seek_window} so all read_state/callback locking stays
co-located with the fields. No behavior change.
